Refactored Component.js in HistoryPage
